+++ spec/spec_helper.rb
@@ -9,6 +9,9 @@ require "missive"
 require "webmock/rspec"
 require "concurrent"
 
+# Require support files
+Dir[File.expand_path("support/**/*.rb", __dir__)].each { |f| require f }
+
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"